代码

+++ sgave-mall-inventory/src/main/java/com/sgave/mall/sgavemallinventory/mapper/InventoryMapper.java
@@ -11,7 +11,7 @@ import org.apache.ibatis.annotations.Param;
  */
 public interface InventoryMapper extends BaseMapper<Inventory> {
 
-    int addStock(@Param("id") Long id, @Param("num") Short num);
-    int reduceStock(@Param("id") Integer id, @Param("num") Short num);
+    int addStock(@Param("id") Long id, @Param("num") Integer num);
+    int reduceStock(@Param("id") Integer id, @Param("num") Integer num);
 
 }
